gltf: Move mesh processing out of gltfpack.cpp
This is a bit messier than animations since we need to move various mesh
analysis/filtering functions as well.
